+++ tests/conftest.py
@@ -9,23 +9,17 @@ def test_env():
     load_dotenv()
 
     blobstore = json.loads(os.getenv('blobstore'))
-    credentials = json.loads(os.getenv('blobstore'))['credentials']
-    mapping = {
-        "blob_endpoint": 'endpoint', 
-        "blob_accessKey": "accessKey",
-        "blob_secretKey": "secretKey"}
-    for k, v in mapping.items():
-        os.environ[k] = credentials[v]
-
-    mapping = {
-        'bucket_name': 'bucket_name',
-        'blob_record_id': 'blob_record_id'
-    }
-    for k, v in mapping.items():
-        os.environ[k] = blobstore[v]
-        
-    os.environ['blob_accessKey'] = str(base64.b64encode(bytes(os.environ['blob_accessKey'], 'utf-8')), 'utf-8')
-    os.environ['blob_secretKey'] = str(base64.b64encode(bytes(os.environ['blob_secretKey'], 'utf-8')), 'utf-8')
+    if 'credentials' in blobstore:
+        credentials = json.loads(os.getenv('blobstore'))['credentials']
+        mapping = {
+            "blob_endpoint":  'endpoint', 
+            "blob_accessKey": "accessKey",
+            "blob_secretKey": "secretKey"}
+        for k, v in mapping.items():
+            os.environ[k] = credentials[v]
+            
+        os.environ['blob_accessKey'] = str(base64.b64encode(bytes(os.environ['blob_accessKey'], 'utf-8')), 'utf-8')
+        os.environ['blob_secretKey'] = str(base64.b64encode(bytes(os.environ['blob_secretKey'], 'utf-8')), 'utf-8')
 
 @pytest.fixture(scope="session")
 def test_param_token():